Distinguish between PointerWidth and Size
